gym_backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email validation
email-validator==2.2.0

# Development and testing
pytest==8.3.5
httpx==0.28.1
flake8==7.2.0
